% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{assemble_element}
\alias{assemble_element}
\title{Copy and adjust tree and state files to create an element}
\usage{
assemble_element(treelabel, traitlabel)
}
\arguments{
\item{treelabel}{Label of Trees to use}

\item{traitlabel}{Label of Traits to use}
}
\value{
Boolean indicating success or failure
}
\description{
The end result is an equal number of name-standardized treefiles and
statefiles.
}
