% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_statefile}
\alias{check_statefile}
\title{Check a trait file submitted to PhyCoMB}
\usage{
check_statefile(statefile)
}
\arguments{
\item{statefile}{The name of the trait file to be checked}
}
\value{
TRUE if the check passes, FALSE if it doesn't
}
\description{
This check function is run on each new trait file submitted to PhyCoMB.  It
expects the file to contain one line per tip as "tiplabel,tipstate", i.e.,
CSV format without column or row names.  Each state should be 0 or 1 (but do
we want to allow NA in order to test with missing data?).
}
