% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_tree_trait}
\alias{check_tree_trait}
\title{Check a tree-trait combination submitted to PhyCoMB}
\usage{
check_tree_trait(treefile, statefile)
}
\arguments{
\item{treefile}{The name of the tree file against which it will be checked}

\item{statefile}{The name of the trait file to be checked}
}
\value{
TRUE if the check passes, FALSE if it doesn't
}
\description{
This check function is run on each item in a TestElement submitted to
PhyCoMB.  It expects a perfect match between the tip labels in the tree and
in the trait matrix.
}
