% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_treefile}
\alias{check_treefile}
\title{Check a tree file submitted to PhyCoMB}
\usage{
check_treefile(treefile)
}
\arguments{
\item{treefile}{The name of the file to be checked}
}
\value{
TRUE if the check passes, FALSE if it doesn't
}
\description{
This check function is run on each new tree file submitted to PhyCoMB.
It expects the file to contain exactly one line, which is a Newick string.
}
