% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make.labels}
\alias{make.labels}
\title{Make random unique labels}
\usage{
make.labels(num = 100, len = 3, seed = NULL, file = NULL)
}
\arguments{
\item{num}{The number of labels to create}

\item{len}{The length of each label}

\item{seed}{Passed to set.seed()}

\item{file}{The name of the file to which to write the labels}
}
\value{
A vector of labels
}
\description{
Generating random labels keeps testing more "blind" than numbering things in order.
}
