% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{multi.to.binary.states}
\alias{multi.to.binary.states}
\title{Convert from multi-state to binary}
\usage{
multi.to.binary.states(phy, n.states)
}
\arguments{
\item{phy}{A tree (phylo, with tip.state attribute)}

\item{n.states}{The number of states before this conversion}
}
\value{
The same tree, with tip states modified
}
\description{
Convert from multi-state to binary
}
