% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{name_statefile}
\alias{name_statefile}
\title{Create filename for states}
\usage{
name_statefile(Z)
}
\arguments{
\item{Z}{The identifier (probably created with \link{zfill})}
}
\value{
A string: s + Z + .csv
}
\description{
Create filename for states
}
