% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{name_treefile}
\alias{name_treefile}
\title{Create filename for tree}
\usage{
name_treefile(Z)
}
\arguments{
\item{Z}{The identifier (probably created with \link{zfill})}
}
\value{
A string: t + Z + .tre
}
\description{
Create filename for tree
}
