% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{neutral_trait_continuous}
\alias{neutral_trait_continuous}
\title{Simulate a neutral continuous-valued trait on a given tree}
\usage{
neutral_trait_continuous(phy, rateval, discretize = TRUE)
}
\arguments{
\item{phy}{A tree}

\item{rateval}{The rate for Brownian motion}

\item{discretize}{Whether to convert the trait to binary}
}
\value{
The tree, with the simulated states as tip.state attribute
}
\description{
Simulate a neutral continuous-valued trait on a given tree
}
