% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{neutral_trait_discrete}
\alias{neutral_trait_discrete}
\title{Simulate a neutral binary trait on a given tree}
\usage{
neutral_trait_discrete(phy, qval, req)
}
\arguments{
\item{phy}{A tree}

\item{qval}{Transition rate parameters; if only one value is given, it is assumed to apply to both q01 and q10}

\item{req}{The requirement for retaining the simulation outcome; either the
minimum proportion of tips in each state (if 0 < req < 0.5) or the minimum
number of tips in each state (if req >= 1)}
}
\value{
Vector of tip states
}
\description{
Simulate a neutral binary trait on a given tree
}
