% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{read.tree.states}
\alias{read.tree.states}
\title{Read in tree file and states file}
\usage{
read.tree.states(Z)
}
\arguments{
\item{Z}{The identifier (probably created with \link{zfill})}
}
\value{
A phylo object, with states in the tip.state attribute
}
\description{
Read in tree file and states file
}
