% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{require.both.states}
\alias{require.both.states}
\title{Require at least N tips of both states}
\usage{
require.both.states(phy.all, N = 1)
}
\arguments{
\item{phy.all}{A list of trees (each with tip.state attribute)}

\item{N}{Require at least N tips in each state}
}
\value{
The subset of phy.all that meets the requirement
}
\description{
Require at least N tips of both states
}
