% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeutil.R
\name{scale.root.age}
\alias{scale.root.age}
\title{Rescale the tree to specified root age}
\usage{
\method{scale}{root.age}(phy, age = 1)
}
\arguments{
\item{phy}{The tree to rescale}

\item{age}{The desired root age}
}
\value{
The rescaled tree
}
\description{
Rescale the tree to specified root age
}
