% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeutil.R
\name{scale_root_age}
\alias{scale_root_age}
\title{Rescale the tree to specified root age}
\usage{
scale_root_age(phy, age = 1)
}
\arguments{
\item{phy}{The tree to rescale}

\item{age}{The desired root age}
}
\value{
The rescaled tree
}
\description{
Rescale the tree to specified root age
}
