% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeutil.R
\name{wipe_tree}
\alias{wipe_tree}
\title{Remove identifying information from a tree}
\usage{
wipe_tree(x, z)
}
\arguments{
\item{x}{The number of the tree (e.g., 14)}

\item{z}{The padded length of the tree's label (e.g., 3 if it's called 014)}
}
\description{
Two files are written: one containing the tree, one containing the tip
states.  In these files, the tip names are wiped of identifying information.
}
