% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeio.R
\name{write.trees}
\alias{write.trees}
\title{Write files for each tree}
\usage{
write.trees(phy.all, z = 3, do.tree = TRUE, do.states = TRUE)
}
\arguments{
\item{phy.all}{A list of trees}

\item{z}{The padded length of the tree's label (e.g., 3 if it's called 014)}

\item{do.tree}{Whether to create a file for each tree}

\item{do.states}{Whether to create a file for each set of tip states}
}
\description{
Two files are written: one containing the tree (tNNN.tre), one containing
the tip states (sNNN.csv).
}
