% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{zfill}
\alias{zfill}
\title{Fill with leading zeros to get an n-digit character string}
\usage{
zfill(x, z)
}
\arguments{
\item{x}{The input string, which is to be padded}

\item{z}{The desired total length of the output string}
}
\value{
The string x preceded by (z-x) zeros.  Or if x is too long to pad, it is returned unchanged.
}
\description{
code is from http://tolstoy.newcastle.edu.au/R/help/00a/1076.html
}
